+++ src/components/Power.svelte
@@ -1,14 +1,22 @@
 <script>
-
+	
 </script>
 
 <main>
-
+	<div id="lock"></div>
 </main>
 
 <style>
 	main {
 		background-color: #2f233f;
+		position: relative;
+	}
+	#lock {
+		background-image: url("../assets/lock.svg");
+		position: absolute;
+		left: 50%; top: 50%;
+		transform: translate(-50%, -50%);
+		width: 5rem; height: 5rem;
 	}
 	/* Button: #bb86fc */
 </style>